--- src/app/EduSchoolBackOffice/Components/gestionPedagogique/classe/eleveClasse/eleve-classe/eleve-classe.component.css
@@ -1,36 +1,0 @@
-.title-and-button {
-  display: flex;
-  justify-content: space-between;
-}
-
-.title {
-  margin-right: auto; /* Aligne le titre à gauche en poussant le bouton vers la droite */
-}
-#btnAdd{
-  background-color: #06BBCC;
-  color:white
-}
-
-.popup {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 300px;
-  background-color: white;
-  box-shadow: 0 5px 15px rgba(0,0,0,0.3);
-  z-index: 1000;
-}
-
-.popup-content {
-  padding: 20px;
-  text-align: center;
-}
-
-.close-btn {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  cursor: pointer;
-}
-
