--- src/app/EduSchoolBackOffice/Components/gestionPedagogique/classe/eleveClasse/eleve-classe/eleve-classe.component.html
@@ -1,43 +1,0 @@
-
-<div class="content">
-
-  <div class="container">
-    <div class="container">
-      <div class="title-and-button">
-        <div  routerLink="../detail-classe">
-          <mat-icon>arrow_back_ios</mat-icon>
-        </div>
-        <h1 class="title">Listes des Enseignants dans la classe {{classe.nom}} </h1>
-        <button id="btnAdd" type="button" class="btn btn-info" (click)="isPopupOpen=true">Affecter
-          éléve</button>
-      </div>
-    </div>
-    <div class="table-responsive">
-    <app-table [tableData]="elevesClasseList"
-              [tableColumns]="elevesClasseTableColumns"
-              [isFilterable]="true"
-              [isPageable]="true"
-              [isDetailIcon]="false"
-              [paginationSizes]="[2,4,6]"
-              [defaultPageSize]="4"
-              [rowActionEditIcon]="'edit'"
-              [rowActionDeleteIcon]="'delete'"
-              [rowActionDetailIcon]="'remove_red_eye'"
-              (rowDeleteAction)="desafecterEleveClasse($event)">
-</app-table>
- </div>
-
-
-
-  </div>
-
-  <div class="popup" *ngIf="isPopupOpen">
-    <div class="popup-content">
-      <span class="close-btn" (click)="closePopup()">&times;</span>
-      <p>Enter some text:</p>
-      <input type="text" />
-      <button (click)="affecterEleve()">Submit</button>
-    </div>
-  </div>
-
-</div>
