+++ src/app/EduSchoolBackOffice/Components/gestionPedagogique/classe/enseignant-classe/enseignant-classe.component.html
@@ -1,0 +1,84 @@
+<div class="container" style="margin-top:50px ;">
+  <div class="row">
+
+
+      <div class="col-md-7">
+        <div class="container">
+          <div class="title-and-button">
+            <div  routerLink="../detail-classe">
+              <mat-icon>arrow_back_ios</mat-icon>
+            </div>
+            <h1 class="title">Listes des Enseignants dans la classe <span class="fw-bold">{{classe.nom}}</span> </h1>
+          </div>
+          <br>
+          <br>
+        </div>
+            <!-- Table -->
+            <table class="table">
+
+              <!-- En-tête de la table -->
+              <ng-container>
+                <th class="table-active">Enseignant</th>
+                <th class="table-active">Matiere Enseignés</th>
+                <th class="table-active" colspan="3">Actions</th>
+                <tr *ngFor="let element of enseignantsClasseList">
+                  <td>
+                    {{element["enseignant"]["nom"]}}
+                  </td>
+                  <td>
+                    <span *ngFor="let e of element['matiereEnseignees']">{{ e['nom'] }},</span>
+                  </td>
+                  <td (click)="editRow(element)">
+                    <button mat-button>
+                      <mat-icon>edit</mat-icon>
+                    </button>
+                  </td>
+                  <td (click)="deleteRow(element)">
+                    <button mat-button>
+                      <mat-icon>delete</mat-icon>
+                    </button>
+                  </td>
+                </tr>
+              </ng-container>
+
+            </table>
+
+      </div>
+      <div class="col-md-5"  >
+        <form  [formGroup]="clEnsForm">
+
+          <h2 class="title">Affecter un enseignant </h2>
+
+          <div class="form-row">
+            <div class="col-md-12">
+                <label for="validationCustom04" class="form-label" >Identifiant de l'enseignant <span style="color: red;">*</span></label>
+              <input type="text" class="form-control" formControlName="idEns" >
+            <div *ngIf="validateError('idEns','required') && submitted"  >
+              <div  class="alert alert-danger">Identifiant est Obligatoire </div>
+            </div>
+          </div>
+          </div>
+          <div class="form-row">
+            <div class="col-md-12">
+                <label for="validationCustom04" class="form-label" >Nom de la classe <span style="color: red;">*</span></label>
+                <select class="form-select" required formControlName="matiere">
+                  <option value="-- Sélectionner une matiere --"  selected>-- Sélectionner une matiere --</option>
+                  <option *ngFor="let n of matieres" [value]="n.nom">{{n.nom}}</option>
+                </select>
+                <div *ngIf="clEnsForm.get('matiere')?.value==='-- Sélectionner une matiere --' && submitted"  >
+                  <div  class="alert alert-danger">selectionner une matiére</div>
+                </div>
+            </div>
+        </div>
+          <div class="form-group">
+              <button type="submit" class="btn btn-block create-account" [disabled]="clEnsForm.invalid" (click)="affecterEns()">Affecter</button>
+          </div>
+      </form>
+      </div>
+  </div>
+
+
+
+
+
+
