--- src/app/EduSchoolBackOffice/Components/gestionPedagogique/classe/enseignantClasse/affecter-enseignant/affecter-enseignant.component.css
@@ -1,46 +1,0 @@
-body{
-  background-color: #dee9ff;
-}
-
-.registration-form{
-padding: 50px 0;
-}
-
-.registration-form form{
-  background-color: #fff;
-  max-width: 600px;
-  margin: auto;
-  padding: 50px 70px;
-  border-top-left-radius: 30px;
-  border-top-right-radius: 30px;
-  box-shadow: 0px 2px 10px rgba(0, 0, 0, 0.075);
-}
-
-.registration-form .form-icon{
-text-align: center;
-  border-radius: 50%;
-  margin: auto;
-  margin-bottom: 50px;
-  line-height: 100px;
-}
-
-
-
-.registration-form .create-account{
-  border-radius: 30px;
-  padding: 10px 20px;
-  font-size: 18px;
-  font-weight: bold;
-  background-color:#06BBCC;
-  border: none;
-  color: white;
-  margin-top: 20px;
-}
-
-@media (max-width: 576px) {
-  .registration-form form{
-      padding: 50px 20px;
-  }
-
-
-}
