--- src/app/EduSchoolBackOffice/Components/gestionPedagogique/classe/enseignantClasse/affecter-enseignant/affecter-enseignant.component.html
@@ -1,36 +1,0 @@
-<div  routerLink="../enseignant-classe">
-  <mat-icon>arrow_back_ios</mat-icon>
-</div>
-<div class="registration-form">
-  <form  [formGroup]="clEnsForm">
-
-      <div class="form-icon">
-          <img src="assets/img/teacher.png">
-      </div>
-      <div class="form-row">
-        <div class="col-md-12">
-            <label for="validationCustom04" class="form-label" >Identifiant de l'enseignant <span style="color: red;">*</span></label>
-          <input type="text" class="form-control" formControlName="idEns" >
-        <div *ngIf="validateError('idEns','required') && submitted"  >
-          <div  class="alert alert-danger">Identifiant est Obligatoire </div>
-        </div>
-      </div>
-      </div>
-      <div class="form-row">
-        <div class="col-md-12">
-            <label for="validationCustom04" class="form-label" >Nom de la classe <span style="color: red;">*</span></label>
-            <select class="form-select" required formControlName="matiere">
-              <option value="-- Sélectionner une matiere --"  selected>-- Sélectionner une matiere --</option>
-              <option *ngFor="let n of matieres" [value]="n.nom">{{n.nom}}</option>
-            </select>
-            <div *ngIf="clEnsForm.get('matiere')?.value==='-- Sélectionner une matiere --' && submitted"  >
-              <div  class="alert alert-danger">selectionner une matiére</div>
-            </div>
-        </div>
-    </div>
-      <div class="form-group">
-          <button type="submit" class="btn btn-block create-account" [disabled]="clEnsForm.invalid" (click)="affecterEns()">Affecter</button>
-      </div>
-  </form>
-
-</div>
