--- src/app/EduSchoolBackOffice/Components/gestionPedagogique/classe/enseignantClasse/enseignant-classe/enseignant-classe.component.css
@@ -1,20 +1,0 @@
-.title-and-button {
-  display: flex;
-  justify-content: space-between;
-}
-
-.title {
-  margin-right: auto; /* Aligne le titre à gauche en poussant le bouton vers la droite */
-}
-#btnAdd{
-  background-color: #06BBCC;
-  color:white
-}
-.transparent-table {
-  background-color: transparent;
-}
-tr{
-  background-color: white;
-}
-
-
