--- src/app/EduSchoolBackOffice/Components/gestionPedagogique/classe/enseignantClasse/enseignant-classe/enseignant-classe.component.html
@@ -1,51 +1,0 @@
-<div class="content">
-
-    <div class="container">
-      <div class="title-and-button">
-        <div  routerLink="../detail-classe">
-          <mat-icon>arrow_back_ios</mat-icon>
-        </div>
-        <h1 class="title">Listes des Enseignants dans la classe {{classe.nom}} </h1>
-        <button id="btnAdd" type="button" class="btn btn-info" (click)="goToAffecterClasse()">Affecter
-          enseignant</button>
-      </div>
-      <br>
-      <br>
-    </div>
-        <!-- Table -->
-        <table class="table">
-
-          <!-- En-tête de la table -->
-          <ng-container>
-            <th class="table-active">Enseignant</th>
-            <th class="table-active">Matiere Enseignés</th>
-            <th class="table-active" colspan="3">Actions</th>
-            <tr *ngFor="let element of enseignantsClasseList">
-              <td>
-                {{element["enseignant"]["nom"]}}
-              </td>
-              <td>
-                <span *ngFor="let e of element['matiereEnseignees']">{{ e['nom'] }},</span>
-              </td>
-              <td (click)="editRow(element)">
-                <button mat-button>
-                  <mat-icon>edit</mat-icon>
-                </button>
-              </td>
-              <td (click)="deleteRow(element)">
-                <button mat-button>
-                  <mat-icon>delete</mat-icon>
-                </button>
-              </td>
-            </tr>
-          </ng-container>
-
-        </table>
-
-
-
-
-
-
-  </div>
-
